Refactor materials into client instead of common
