+++ src/Client/ClientCore/Rendering/Materials/MaterialInstaller.cs
@@ -19,7 +19,7 @@ using Castle.MicroKernel.Registration;
 using Castle.MicroKernel.SubSystems.Configuration;
 using Castle.Windsor;
 
-namespace Sovereign.EngineCore.World.Materials;
+namespace Sovereign.ClientCore.Rendering.Materials;
 
 /// <summary>
 ///     IoC installer for materials support classes.
