+++ src/Client/ClientCore/Rendering/Materials/MaterialManager.cs
@@ -19,11 +19,12 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Castle.Core.Logging;
+using Sovereign.ClientCore.Rendering.Sprites.TileSprites;
 using Sovereign.EngineCore.Logging;
 using Sovereign.EngineCore.Main;
 using Sovereign.EngineCore.Resources;
 
-namespace Sovereign.EngineCore.World.Materials;
+namespace Sovereign.ClientCore.Rendering.Materials;
 
 /// <summary>
 ///     Responsible for managing materials.
@@ -49,14 +50,20 @@ public class MaterialManager
     /// </summary>
     private readonly IResourcePathBuilder pathBuilder;
 
+    private readonly TileSpriteManager tileSpriteManager;
+
     public MaterialManager(MaterialDefinitionsLoader loader, IResourcePathBuilder pathBuilder,
-        ILogger logger, IErrorHandler errorHandler)
+        ILogger logger, IErrorHandler errorHandler, TileSpriteManager tileSpriteManager)
     {
         /* Set dependencies. */
         this.loader = loader;
         this.pathBuilder = pathBuilder;
         this.logger = logger;
         this.errorHandler = errorHandler;
+        this.tileSpriteManager = tileSpriteManager;
+
+        tileSpriteManager.OnTileSpriteAdded += OnTileSpriteAdded;
+        tileSpriteManager.OnTileSpriteRemoved += OnTileSpriteRemoved;
     }
 
     /// <summary>
@@ -132,5 +139,21 @@ public class MaterialManager
             .Append(airMat)
             .OrderBy(material => material.Id)
             .ToList();
+    }
+
+    /// <summary>
+    ///     Called when a new tile sprite is added.
+    /// </summary>
+    /// <param name="tileSpriteId">Added tile sprite ID.</param>
+    private void OnTileSpriteAdded(int tileSpriteId)
+    {
+    }
+
+    /// <summary>
+    ///     Called when a tile sprite is removed.
+    /// </summary>
+    /// <param name="tileSpriteId">Removed tile sprite ID.</param>
+    private void OnTileSpriteRemoved(int tileSpriteId)
+    {
     }
 }
