+++ src/Client/ClientCore/Rendering/RenderingResourceManager.cs
@@ -20,7 +20,6 @@ using Sovereign.ClientCore.Rendering.Sprites;
 using Sovereign.ClientCore.Rendering.Sprites.AnimatedSprites;
 using Sovereign.ClientCore.Rendering.Sprites.Atlas;
 using Sovereign.ClientCore.Rendering.Sprites.TileSprites;
-using Sovereign.EngineCore.World.Materials;
 
 namespace Sovereign.ClientCore.Rendering;
