+++ src/Client/ClientCore/Rendering/Scenes/Game/World/WorldTileSpriteSequencer.cs
@@ -16,13 +16,13 @@
  */
 
 using System.Collections.Generic;
+using Sovereign.ClientCore.Rendering.Materials;
 using Sovereign.ClientCore.Rendering.Sprites.TileSprites;
 using Sovereign.ClientCore.Systems.Block.Caches;
 using Sovereign.EngineCore.Components;
 using Sovereign.EngineCore.Components.Types;
 using Sovereign.EngineCore.Systems.Block.Components;
 using Sovereign.EngineCore.Systems.Block.Components.Indexers;
-using Sovereign.EngineCore.World.Materials;
 
 namespace Sovereign.ClientCore.Rendering.Scenes.Game.World;
