+++ src/Client/ClientCore/Systems/Block/Caches/BlockAnimatedSpriteCache.cs
@@ -22,13 +22,13 @@ using System.Linq;
 using System.Numerics;
 using System.Threading.Tasks;
 using Castle.Core.Logging;
+using Sovereign.ClientCore.Rendering.Materials;
 using Sovereign.ClientCore.Rendering.Sprites.TileSprites;
 using Sovereign.EngineCore.Components;
 using Sovereign.EngineCore.Components.Indexers;
 using Sovereign.EngineCore.Entities;
 using Sovereign.EngineCore.Systems.Block.Components;
 using Sovereign.EngineCore.Systems.Block.Components.Indexers;
-using Sovereign.EngineCore.World.Materials;
 using Sovereign.EngineUtil.Collections;
 
 namespace Sovereign.ClientCore.Systems.Block.Caches;
