+++ src/Common/EngineCore/World/MaterialConstants.cs
@@ -1,0 +1,28 @@
+// Sovereign Engine
+// Copyright (c) 2024 opticfluorine
+// 
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+// 
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+// 
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+namespace Sovereign.EngineCore.World;
+
+/// <summary>
+///     Common constants related to materials.
+/// </summary>
+public static class MaterialConstants
+{
+    /// <summary>
+    ///     Material ID of air.
+    /// </summary>
+    public const int Air = 0;
+}
